+++ apps/E2E/src/FocusTrapZone/pages/FocusTrapZonePageObject.win.ts
@@ -6,13 +6,6 @@ import {
 import { BasePage, By } from '../../common/BasePage';
 
 class FocusTrapZonePageObject extends BasePage {
-  // OVERRIDE: We use isExisting() here instead of isDisplayed() because FocusTrapZone does not have any UI to it, it's simply
-  // a wrapper that adds keyboard focus functionality
-  async waitForPrimaryElementDisplayed(timeout?: number): Promise<void> {
-    const errorMsg = 'The FocusTrapZone UI Element did not load correctly. Please see logs.';
-    await this.waitForCondition(async () => await (await this._primaryComponent).isExisting(), errorMsg, timeout);
-  }
-
   /*****************************************/
   /**************** Getters ****************/
   /*****************************************/