+++ apps/E2E/src/Menu/specs/Menu.spec.win.ts
@@ -24,8 +24,9 @@ describe('Menu Testing Initialization', function () {
 describe('Menu Accessibility Testing', () => {
   beforeAll(async () => {
     await MenuPageObject.scrollToTestElement();
-    await MenuPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
     await MenuPageObject.openMenu();
+
+    await MenuPageObject.resetTest();
   });
 
   it('Menu - Validate accessibilityRole of menu item is correct', async () => {
@@ -55,7 +56,7 @@ describe('Menu Functional Testing', () => {
   /* Scrolls and waits for the Menu to be visible on the Test Page */
   beforeEach(async () => {
     await MenuPageObject.scrollToTestElement();
-    await MenuPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
+
     await MenuPageObject.resetTest();
   });