+++ apps/E2E/src/RadioGroup/specs/RadioGroup.spec.win.ts
@@ -30,7 +30,6 @@ describe('RadioGroup/RadioButton Accessibility Testing', () => {
   /* Scrolls and waits for the RadioGroup to be visible on the Test Page */
   beforeEach(async () => {
     await RadioGroupPageObject.scrollToTestElement();
-    await RadioGroupPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
   });
 
   it("Validate RadioGroup's accessibilityRole is correct", async () => {
@@ -70,7 +69,6 @@ describe('RadioGroup Functional Testing', async () => {
   /* This resets the RadioGroup state by clicking/selecting the 1st RadioButton in the RadioGroup */
   beforeEach(async () => {
     await RadioGroupPageObject.scrollToTestElement();
-    await RadioGroupPageObject.waitForPrimaryElementDisplayed(PAGE_TIMEOUT);
 
     await RadioGroupPageObject.resetRadioGroupSelection();
   });