+++ apps/fluent-tester/src/TestComponents/TextExperimental/TextExperimentalTest.tsx
@@ -33,7 +33,7 @@ const textSections: TestSection[] = [
 
 export const TextExperimentalTest: React.FunctionComponent = () => {
   const status: PlatformStatus = {
-    win32Status: 'Experimental',
+    win32Status: 'Beta',
     uwpStatus: 'Experimental',
     iosStatus: 'Experimental',
     macosStatus: 'Experimental',